enhancement: IconSelect new theme
